+++ src/gallium/state_trackers/nine/Makefile.sources
@@ -5,6 +5,8 @@ C_SOURCES := \
 	authenticatedchannel9.h \
 	basetexture9.c \
 	basetexture9.h \
+	buffer9.c \
+	buffer9.h \
 	cryptosession9.c \
 	cryptosession9.h \
 	cubetexture9.c \
