+++ src/gallium/state_trackers/nine/buffer9.h
@@ -1,5 +1,6 @@
 /*
  * Copyright 2011 Joakim Sindholt <opensource@zhasha.com>
+ * Copyright 2015 Patrick Rudolph <siro@das-labor.org>
  *
  * Permission is hereby granted, free of charge, to any person obtaining a
  * copy of this software and associated documentation files (the "Software"),
@@ -20,60 +21,53 @@
  * OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE
  * USE OR OTHER DEALINGS IN THE SOFTWARE. */
 
-#ifndef _NINE_VERTEXBUFFER9_H_
-#define _NINE_VERTEXBUFFER9_H_
+#ifndef _NINE_BUFFER9_H_
+#define _NINE_BUFFER9_H_
+
 #include "resource9.h"
-#include "buffer9.h"
 
 struct pipe_screen;
 struct pipe_context;
 struct pipe_transfer;
 
-struct NineVertexBuffer9
+struct NineBuffer9
 {
-    struct NineBuffer9 base;
+    struct NineResource9 base;
 
     /* G3D */
     struct pipe_context *pipe;
-    D3DVERTEXBUFFER_DESC desc;
+    struct pipe_transfer **maps;
+    int nmaps, maxmaps;
+    UINT size;
 };
-static inline struct NineVertexBuffer9 *
-NineVertexBuffer9( void *data )
+static inline struct NineBuffer9 *
+NineBuffer9( void *data )
 {
-    return (struct NineVertexBuffer9 *)data;
+    return (struct NineBuffer9 *)data;
 }
 
 HRESULT
-NineVertexBuffer9_new( struct NineDevice9 *pDevice,
-                       D3DVERTEXBUFFER_DESC *pDesc,
-                       struct NineVertexBuffer9 **ppOut );
-
-HRESULT
-NineVertexBuffer9_ctor( struct NineVertexBuffer9 *This,
+NineBuffer9_ctor( struct NineBuffer9 *This,
                         struct NineUnknownParams *pParams,
-                        D3DVERTEXBUFFER_DESC *pDesc );
+                        D3DRESOURCETYPE Type,
+                        DWORD Usage,
+                        UINT Size,
+                        D3DPOOL Pool );
 
 void
-NineVertexBuffer9_dtor( struct NineVertexBuffer9 *This );
-/*** Nine private ***/
+NineBuffer9_dtor( struct NineBuffer9 *This );
 
 struct pipe_resource *
-NineVertexBuffer9_GetResource( struct NineVertexBuffer9 *This );
-
-/*** Direct3D public ***/
+NineBuffer9_GetResource( struct NineBuffer9 *This );
 
 HRESULT WINAPI
-NineVertexBuffer9_Lock( struct NineVertexBuffer9 *This,
+NineBuffer9_Lock( struct NineBuffer9 *This,
                         UINT OffsetToLock,
                         UINT SizeToLock,
                         void **ppbData,
                         DWORD Flags );
 
 HRESULT WINAPI
-NineVertexBuffer9_Unlock( struct NineVertexBuffer9 *This );
+NineBuffer9_Unlock( struct NineBuffer9 *This );
 
-HRESULT WINAPI
-NineVertexBuffer9_GetDesc( struct NineVertexBuffer9 *This,
-                           D3DVERTEXBUFFER_DESC *pDesc );
-
-#endif /* _NINE_VERTEXBUFFER9_H_ */
+#endif /* _NINE_BUFFER9_H_ */
